rename api sub-app to resolver_api
