--- api/api/__init__.py
@@ -1,3 +1,0 @@
-from api.api import views
-
-__all__ = ["views"]
