--- api/api/resources/__init__.py
@@ -1,11 +1,0 @@
-from api.api.resources.user import UserResource, UserList
-from api.api.resources.compound import CompoundResource, CompoundList, CompoundSearch
-
-
-__all__ = [
-    "UserResource",
-    "UserList",
-    "CompoundResource",
-    "CompoundList",
-    "CompoundSearch",
-]
