--- api/api/schemas/__init__.py
@@ -1,4 +1,0 @@
-from api.api.schemas.user import UserSchema
-from api.api.schemas.compound import CompoundSchema
-
-__all__ = ["UserSchema", "CompoundSchema"]
