+++ api/manage.py
@@ -30,3 +30,14 @@ def init():
 
 if __name__ == "__main__":
     cli()
+
+
+@cli.command("load_seed_data")
+def load_seed_data():
+    # https://chem.nlm.nih.gov/chemidplus/rn/1050-79-9
+    c1 = Compound(
+        id="DTXCID302000003",
+        identifiers='{ "preferred_name":"Moperone","casrn":"1050-79-9","inchikey": "AGAHNABIDCTLHW-UHFFFAOYSA-N", "casalts":[{"casalt":"0001050799","weight":0.5},{"casalt":"1050799","weight":0.5}],"synonyms": [{"synonym": "Meperon","weight": 0.75},{"synonym": "Methylperidol","weight": 0.5}]}',
+    )
+    db.session.add(c1)
+    db.session.commit()
