+++ api/resolver_api/__init__.py
@@ -1,0 +1,3 @@
+from api.resolver_api import views
+
+__all__ = ["views"]
