+++ api/resolver_api/resources/__init__.py
@@ -1,0 +1,15 @@
+from api.resolver_api.resources.user import UserResource, UserList
+from api.resolver_api.resources.compound import (
+    CompoundResource,
+    CompoundList,
+    CompoundSearch,
+)
+
+
+__all__ = [
+    "UserResource",
+    "UserList",
+    "CompoundResource",
+    "CompoundList",
+    "CompoundSearch",
+]
