+++ api/resolver_api/resources/user.py
@@ -1,7 +1,7 @@
 from flask import request
 from flask_restful import Resource
 from flask_jwt_extended import jwt_required
-from api.api.schemas import UserSchema
+from api.resolver_api.schemas import UserSchema
 from api.models import User
 from api.extensions import db
 from api.commons.pagination import paginate
