+++ api/resolver_api/schemas/__init__.py
@@ -1,0 +1,4 @@
+from api.resolver_api.schemas.user import UserSchema
+from api.resolver_api.schemas.compound import CompoundSchema
+
+__all__ = ["UserSchema", "CompoundSchema"]
