+++ template.env
@@ -14,6 +14,6 @@ SQL_USER=postgres
 # The redundancy should be tidied up
 
 # The server is using 5431 to avoid conflicts with other apps
-DATABASE_URL=postgresql://postgres:postgres@127.0.0.1:5431/resolver
+DATABASE_URI=postgresql://postgres:postgres@127.0.0.1:5431/resolver
